refactor(phase-19)(pr-11): decouple moduleStoreCleanup from gameSessionStore
PR-8 (#91) code review HIGH severity finding 해소:
moduleStoreFactory.ts ↔ gameSessionStore.ts 순환 import (top-level).

Before:
  moduleStoreFactory.ts  → gameSessionStore.ts (useGameSessionStore import)
  gameSessionStore.ts    → moduleStoreFactory.ts (clearBySessionId import)
  = 양방향 top-level cycle (ESM live binding에 의존)

After (3-file split):
  moduleStoreFactory.ts   : core Map + getModuleStore + selectors (no store imports)
  moduleStoreCleanup.ts   : clearBySessionId / clearModuleStores (NEW, factory Map만 import)
  useModuleStore.ts       : React hook (NEW, factory + gameSessionStore import)
  gameSessionStore.ts     : clearBySessionId import 경로 → moduleStoreCleanup

factory는 더 이상 domain store를 참조하지 않는다. 단방향 의존성:
  gameSessionStore → moduleStoreCleanup → moduleStoreFactory
  useModuleStore → { moduleStoreFactory, gameSessionStore }

Backward compat: moduleStoreFactory가 useModuleStore/clearBySessionId/
clearModuleStores를 re-export해 기존 feature 컴포넌트 import 경로 유지.
re-export는 namespace redirection만 수행 (top-level value read 없음) →
runtime cycle 위험 없음.

Verification:
  - pnpm -C apps/web exec tsc --noEmit → 0 errors
  - pnpm -C apps/web test -- --run → 109 files / 1047 tests pass (regression 0)
  - pnpm -C apps/web lint → 0 errors (42 pre-existing warnings 유지)

Files:
  M apps/web/src/stores/moduleStoreFactory.ts            (166 → 135 lines, -31)
  A apps/web/src/stores/moduleStoreCleanup.ts            (49 lines)
  A apps/web/src/stores/useModuleStore.ts                (43 lines)
  M apps/web/src/stores/gameSessionStore.ts              (import path 1 line)
  M apps/web/src/stores/__tests__/moduleStoreFactory.test.ts (import 경로 분리)

공개 API 시그니처 불변 (useModuleStore, getModuleStore, clearBySessionId,
clearModuleStores 그대로). Feature 컴포넌트 수정 0건.
